+++ ui/components/ui/entities/get-provider-logo.tsx
@@ -5,6 +5,7 @@ import {
   AzureProviderBadge,
   GCPProviderBadge,
   GitHubProviderBadge,
+  IacProviderBadge,
   KS8ProviderBadge,
   M365ProviderBadge,
   OracleCloudProviderBadge,
@@ -25,6 +26,8 @@ export const getProviderLogo = (provider: ProviderType) => {
       return <M365ProviderBadge width={35} height={35} />;
     case "github":
       return <GitHubProviderBadge width={35} height={35} />;
+    case "iac":
+      return <IacProviderBadge width={35} height={35} />;
     case "oci":
       return <OracleCloudProviderBadge width={35} height={35} />;
     default:
@@ -46,6 +49,8 @@ export const getProviderName = (provider: ProviderType): string => {
       return "Microsoft 365";
     case "github":
       return "GitHub";
+    case "iac":
+      return "Infrastructure as Code";
     case "oci":
       return "Oracle Cloud Infrastructure";
     default: